refactor(S3Service): abstract S3 logic from ImageService to S3Service
